cmds: subvolume: Add --subvolid argument to subvol_delete
This can be used when a system has a file system mounted from a
subvolume, rather than the root file system, like below:

/
|- @subvol1
|- @subvol2
\- @subvol_default

If only @subvol_default is mounted, we have no path to reach
@subvol1 and @subvol2, thus no way to delete them.
This patch introduces a new argument to specify the subvolume id of the
subvolume/snapshot to be deleted.

If only @subvol_default is mounted, we have no path to reach
@subvol1 and @subvol2, thus no way to delete them. Current subvolume
delete ioctl takes a file handler point as argument, and if
@subvol_default is mounted, we can't reach @subvol1 and @subvol2 from
the same mount point.

This patch introduces a new flag to allow BTRFS_IOC_SNAP_DESTORY_V2
to delete subvolume using subvolid.

Now, we can use this new ioctl specifying the subvolume id and refer to
the same mount point. It's doesn't matter which subvolume was mounted,
since we can reach to the desired one using the subvolume id, and then
delete it.

Now "subvolume delete" receives a new argument --subvolid, which will
contain the subvolume the user wants to delete.

Fixes: #152
Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>