btrfs-progs: mkfs-test: Fix check for truncate command failing
Commit 31f477ee ("btrfs-progs: mkfs-tests: skip test if truncate fails
with EFBIG") tried to detect a failure in truncate command by checking
the $? expecting it to be an errno, when it actually returns 0 or 1.

To fix this test just check if the command failed (returned 1) and look
for the output, skipping the test if the OS cannot create a 6E file.

Fixes: #241

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>