cmds: receive: Drop the xattr hack
The current kernel version was always emitting xattr commands _before_
the chown, and so receive was solving this issue by caching the xattr,
and reapplying it after the chown was executing for he same inode.

The kernel code was fixed, and now the xattr is always sent after the
chown commands, so this hack can be removed.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>